pyproject.toml
[project]
name = "fin-assitant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.4.15",
    "openai>=1.40.0",
    "pandas>=2.0.3",
    "plotly>=5.17.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.0",
    "streamlit>=1.32.0",
]
